requirements.txt
tensorflow-cpu
keras==2.4.3
imutils==0.5.4
numpy==1.19.5
opencv-contrib-python-headless
matplotlib==3.4.2
argparse==1.1
scipy==1.4.1
scikit-learn==0.24.2
pillow==8.2.0
streamlit==0.82.0
streamlit-webrtc==0.20.1
pandas